ARM: i.MX: ocotp: Fix MAC address provider for unaligned addresses
The current algorithm assumes the MAC address starts at a 4 byte
aligned address. Unfortunately this is not always the case. On the
i.MX6sx the MAC Address for the second FEC starts at offset 0x632.
The register space for the fuse map has holes, only the 16 byte
aligned words contain data. This means we have to skip the holes.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>