video: Add simple-panel support
This adds support for simple panels. These are panels which don't need
special handling but can have enable gpios and such. Unlike the Linux
kernel implementation this one is able to understand display-timings
nodes so that it's not necessary to keep a list of all known displays
with their corresponding timings in barebox.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>