serial: stm_serial: Add devicetree support
This adds device tree support for the stm serial driver. This
driver really is a AMBA primecell, so the amba-pl011 could be
used. However, the current code tries to get the apb_pclk before
the clocks are registered, so this does not work. Use the stm
driver instead until a solution is found.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>