ARM: do not specify -static and -pie at the same time
PIE is a form of dynamic linking and thus inherently incompatible
with -static. It worked ok as the current behavior of ld.bfd is
to not respect -static if -pie has been specified.

ld.gold and future versions of ld.bfd will fail to link if both
of those incompatible switches are specified at the same time.

Signed-off-by: Lucas Stach <dev@lynxeye.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>