mtd: m25p80: Fix Kconfig dependencies
The m25p80 driver now depends on MTD_SPI_NOR which is disabled in all
defconfigs, so this effectively disables the m25p80 driver in all
defconfigs. Fix this by selecting MTD_SPI_NOR which is library code
without further dependencies. Also let m25p80 depend on SPI because
it needs the SPI code and won't link without it.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>